% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{cards_container}
\alias{cards_container}
\title{Shiny.semantic card}
\usage{
cards_container(data, column, description, icon, label = NULL)
}
\arguments{
\item{data}{The dataframe}

\item{column}{Reference to the dataframe column}

\item{description}{A description of the column name}

\item{icon}{An icon reference}

\item{label}{A label for the column}
}
\value{
A div
}
\description{
Returns a shiny.semantic card div by a reference of a dataframe
}
