% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_description}
\alias{data_description}
\title{Description of the Marine dataset}
\format{
A data frame with 21 rows and 4 variables:
\describe{
  \item{ID}{Description ID, related to marine dataset column name}
  \item{Description}{A Description about}
  \item{icon}{A tag icon reference }
  \item{label}{A label for the column }
}
}
\usage{
data_description
}
\description{
A dataset that contains a full description of the marine dataset.
It also includes a set of icon and a label name reference to use 
in shiny.
}
\keyword{datasets}
